progressive/stage2/frontend/pyproject.toml
[project]
name = "frontend"
version = "1.0.0"
description = "FastAPI frontend for dice roller with OpenTelemetry instrumentation"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "requests>=2.32.0",
    "prometheus-client>=0.21.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-requests>=0.48b0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
]
